4.23 tehty

+++ osa4/blogilista/controllers/blogs.js
@@ -1,11 +1,16 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const tokenExtractor = require('../utils/middleware').tokenExtractor
+const userExtractor = require('../utils/middleware').userExtractor
 
+//GET all blogs
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
+//GET blog by id
 blogsRouter.get('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id)
   if (blog) {
@@ -15,26 +20,44 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.post('/', async (request, response) => {
+//POST blog
+blogsRouter.post('/', tokenExtractor, userExtractor, async (request, response) => {
   const body = request.body
+  const user = request.user
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes 
+    likes: body.likes,
+    user: user.id
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog.id)
+  await user.save()
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+//DELETE blog
+blogsRouter.delete('/:id', tokenExtractor, userExtractor, async (request, response) => {
+  const userId = request.user.id
+  const blogs = await Blog.findById(request.params.id)
+
+  if(blogs.user.toString() === userId.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end() 
+  } else {
+    response.status(401).json({ error: 'deleting unauthorised'})
+  }
 })
 
-blogsRouter.put('/:id', async (request, response) => {
+//PUT blog
+blogsRouter.put('/:id', tokenExtractor, userExtractor, async (request, response) => {
+  const userId = request.user.id
+  const blogs = await Blog.findById(request.params.id)
   const body = request.body
+
   const blog = {
     title: body.title,
     author: body.author,
@@ -42,8 +65,12 @@ blogsRouter.put('/:id', async (request, response) => {
     likes: body.likes
   }
 
- const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new: true, runValidators: true})
- response.json(updatedBlog)
+  if(blogs.user.toString() === userId.toString()) {
+    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new: true, runValidators: true})
+    response.json(updatedBlog)
+  } else {
+    response.status(401).json({ error: 'updating unauthorised'})
+  }
 })
 
 module.exports = blogsRouter
